Enum forkjoin::ResultReceiver

pub enum ResultReceiver<Ret: Send + Sync> {
    Join(Unique<Ret>, Arc<JoinBarrier<Ret>>),
    Channel(Arc<Mutex<Sender<Ret>>>),
}

Enum describing what to do with results of Tasks and JoinBarriers.

Variants

Join

Algorithm has Summa style and the value should be inserted into a JoinBarrier

Channel

Algorithm has Search style and results should be sent directly to the owner.