Enum forkjoin::SupervisorMsg

pub enum SupervisorMsg<Arg: Send, Ret: Send + Sync> {
    OutOfWork(usize),
    Schedule(Task<Arg, Ret>),
    Shutdown,
}

Messages from ForkPool and WorkerThread to the PoolSupervisor.

Variants

OutOfWork

The WorkerThreads use this to tell the PoolSupervisor they don't have anything to do and that stealing did not give any new Tasks. The argument usize is the id of the WorkerThread.

Schedule

The ForkPool uses this to schedule new tasks on the PoolSupervisor. The PoolSupervisor will later schedule these to a WorkerThread when it see fit.

Shutdown

Message from the ForkPool to the PoolSupervisor to tell it to shutdown.