Enum forkjoin::WorkerMsg

pub enum WorkerMsg<Arg: Send, Ret: Send + Sync> {
    Schedule(Task<Arg, Ret>),
    Steal,
}

Messages from the PoolSupervisor to WorkerThreads

Variants

Schedule

A new Task to be scheduled for execution by the WorkerThread

Steal

Tell the WorkerThread to simply try to steal from the other WorkerThreads