Struct forkjoin::AlgoOnPool

pub struct AlgoOnPool<'a, Arg: 'a + Send, Ret: 'a + Send + Sync> {
    // some fields omitted
}

A handle for a specific Algorithm running on a ForkPool. Acquired from ForkPool::init_algorithm.

Methods

impl<'a, Arg: Send, Ret: Send + Sync> AlgoOnPool<'a, Arg, Ret>

fn schedule(&self, arg: Arg) -> Job<Ret>

Schedule a new computation. Returns instantly with a handle to the computation.

Return value(s) can be read from the returned Job. AlgoStyle::Reduce will only return one message on this channel.

AlgoStyle::Search algorithm might return arbitrary number of messages.