Struct forkjoin::Algorithm

pub struct Algorithm<Arg: Send, Ret: Send + Sync> {
    pub fun: TaskFun<Arg, Ret>,
    pub style: AlgoStyle<Ret>,
}

The representation of a specific algorithm to use the ForkJoin library.

Create one instance of this struct for each algorithm to be executed in ForkJoin.

Fields

fun

A function pointer. The function that will be executed by all the tasks and subtasks.

style

Enum showing the type of algorithm, indicate what should be done with results from subtasks created by forks of this algorithm.