Struct forkjoin::Fork

pub struct Fork<Arg, Ret> {
    pub fun: TaskFun<Arg, Ret>,
    pub args: Vec<Arg>,
    pub join: AlgoStyle<Ret>,
}

Struct describing how a Task want to fork into multiple subtasks.

Fields

fun

A function pointer. The function that will be executed by all the subtasks

args

A list of the arguments to the subtasks. One subtask will be created for each argument.

join

Enum showing the type of algorithm, indicate what should be done with results from subtasks created by this fork.