Struct forkjoin::Job

pub struct Job<Ret> {
    // some fields omitted
}

The handle for a computation. Can be used to fetch results of the computation. Upon drop it will wait for the entire computation to complete if it's still executing. Algorithm termination is detected by the try_recv and recv methods returning a ResultError

Methods

impl<Ret> Job<Ret>

fn try_recv(&self) -> Result<Ret, ResultError>

Attempt to get a result from this Job without blocking. Will return a ResultError if no result is available at the time of call.

fn recv(&self) -> Result<Ret, ResultError>

Block and wait for a result. If a result is available it will return immediately.