Enum forkjoin::AlgoStyle

pub enum AlgoStyle<Ret> {
    Reduce(ReduceStyle<Ret>),
    Search,
}

Enum representing the style of the executed algorithm.

Variants

Reduce

A Reduce style algorithm join together the results of the individual nodes in the problem tree to finally form one result for the entire computation.

Examples of this style include recursively computing fibbonacci numbers and summing binary trees.

Takes a ReduceStyle to indicate what type of join function to use.

Search

A Search style algoritm return their results to the listener directly upon a TaskResult::Done.

Examples of this style include sudoku solvers and nqueens where a node can represent a complete solution.